# openapi.json
{"openapi": "3.1.0", "info": {"title": "FinBot", "description": "Intelig\u00eancia financeira ao seu alcance", "version": "0.1.0"}, "paths": {"/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "description": "Login OAuth2 com username e password.", "operationId": "login_auth_login_post", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_login_auth_login_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/users/": {"post": {"tags": ["users"], "summary": "Create User", "description": "Criar novo usu\u00e1rio.", "operationId": "create_user_users_users__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/users/me": {"get": {"tags": ["users"], "summary": "Read User Me", "description": "Obter usu\u00e1rio atual.", "operationId": "read_user_me_users_users_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserOut"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/finance/finance/accounts/": {"get": {"tags": ["finance"], "summary": "Get Accounts", "description": "Listar todas as contas do usu\u00e1rio", "operationId": "get_accounts_finance_finance_accounts__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Account"}, "type": "array", "title": "Response Get Accounts Finance Finance Accounts  Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["finance"], "summary": "Create Account", "description": "Criar nova conta", "operationId": "create_account_finance_finance_accounts__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountBase"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/finance/finance/transactions/": {"post": {"tags": ["finance"], "summary": "Create Transaction", "description": "Criar nova transa\u00e7\u00e3o", "operationId": "create_transaction_finance_finance_transactions__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionBase"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/finance/finance/bills/": {"post": {"tags": ["finance"], "summary": "Create Bill", "description": "Criar nova conta a pagar", "operationId": "create_bill_finance_finance_bills__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/BillBase"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/finance/finance/bills/pending": {"get": {"tags": ["finance"], "summary": "Get Pending Bills", "description": "Listar contas a pagar pendentes", "operationId": "get_pending_bills_finance_finance_bills_pending_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/finance/finance/export/transactions": {"get": {"tags": ["finance"], "summary": "Export Transactions", "description": "Exporta transa\u00e7\u00f5es em CSV ou Excel", "operationId": "export_transactions_finance_finance_export_transactions_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "start_date", "in": "query", "required": true, "schema": {"type": "string", "format": "date-time", "title": "Start Date"}}, {"name": "end_date", "in": "query", "required": true, "schema": {"type": "string", "format": "date-time", "title": "End Date"}}, {"name": "format", "in": "query", "required": false, "schema": {"type": "string", "default": "csv", "title": "Format"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/finance/finance/categories/": {"get": {"tags": ["finance"], "summary": "Get Categories", "description": "Listar todas as categorias", "operationId": "get_categories_finance_finance_categories__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Category"}, "type": "array", "title": "Response Get Categories Finance Finance Categories  Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["finance"], "summary": "Create Category", "description": "Criar nova categoria", "operationId": "create_category_finance_finance_categories__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CategoryBase"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/finance/finance/categories/{category_id}": {"get": {"tags": ["finance"], "summary": "Get Category", "description": "Obter categoria por ID", "operationId": "get_category_finance_finance_categories__category_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "category_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Category Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Category"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/finance/finance/goals/": {"get": {"tags": ["finance"], "summary": "Get Goals", "description": "Listar todas as metas do usu\u00e1rio", "operationId": "get_goals_finance_finance_goals__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Goal"}, "type": "array", "title": "Response Get Goals Finance Finance Goals  Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["finance"], "summary": "Create Goal", "description": "Criar nova meta", "operationId": "create_goal_finance_finance_goals__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Goal"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/finance/finance/goals/{goal_id}": {"get": {"tags": ["finance"], "summary": "Get Goal", "description": "Obter meta por ID", "operationId": "get_goal_finance_finance_goals__goal_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "goal_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Goal Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Goal"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/finance/finance/goals/{goal_id}/update-amount": {"put": {"tags": ["finance"], "summary": "Update Goal Amount", "description": "Atualizar valor atual da meta", "operationId": "update_goal_amount_finance_finance_goals__goal_id__update_amount_put", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "goal_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Goal Id"}}, {"name": "amount", "in": "query", "required": true, "schema": {"type": "number", "title": "Amount"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/whatsapp/send-message": {"post": {"tags": ["whatsapp"], "summary": "Send Whatsapp Message", "description": "Enviar mensagem via WhatsApp\n\nO n\u00famero deve estar no formato: 11999999999 (DDD + n\u00famero)\nA mensagem pode conter texto livre", "operationId": "send_whatsapp_message_whatsapp_send_message_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessageRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/whatsapp/qr": {"get": {"tags": ["whatsapp"], "summary": "Get Qr Code", "description": "Gera QR Code para conex\u00e3o do WhatsApp", "operationId": "get_qr_code_whatsapp_qr_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/whatsapp/webhook": {"post": {"tags": ["whatsapp"], "summary": "Whatsapp Webhook", "description": "Webhook para receber mensagens do WhatsApp", "operationId": "whatsapp_webhook_whatsapp_webhook_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/whatsapp/test/{command}": {"get": {"tags": ["whatsapp"], "summary": "Test Command", "description": "Testa comandos do bot em uma p\u00e1gina web", "operationId": "test_command_whatsapp_test__command__get", "parameters": [{"name": "command", "in": "path", "required": true, "schema": {"type": "string", "title": "Command"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"summary": "Health Check", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"Account": {"properties": {"name": {"type": "string", "title": "Name"}, "balance": {"type": "number", "title": "Balance", "default": 0.0}, "type": {"type": "string", "title": "Type", "default": "checking"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}, "owner_id": {"type": "integer", "title": "Owner Id"}}, "type": "object", "required": ["name", "owner_id"], "title": "Account"}, "AccountBase": {"properties": {"name": {"type": "string", "title": "Name"}, "balance": {"type": "number", "title": "Balance", "default": 0.0}, "type": {"type": "string", "title": "Type", "default": "checking"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}}, "type": "object", "required": ["name"], "title": "AccountBase"}, "BillBase": {"properties": {"description": {"type": "string", "title": "Description"}, "amount": {"type": "number", "title": "Amount"}, "due_date": {"type": "string", "format": "date-time", "title": "Due Date"}, "is_paid": {"type": "boolean", "title": "Is Paid", "default": false}, "category_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Category Id"}}, "type": "object", "required": ["description", "amount", "due_date"], "title": "BillBase"}, "Body_login_auth_login_post": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "password"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_login_auth_login_post"}, "Category": {"properties": {"name": {"type": "string", "title": "Name"}, "type": {"type": "string", "title": "Type"}, "icon": {"type": "string", "title": "Icon", "default": "\ud83d\udcb0"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}}, "type": "object", "required": ["name", "type"], "title": "Category"}, "CategoryBase": {"properties": {"name": {"type": "string", "title": "Name"}, "type": {"type": "string", "title": "Type"}, "icon": {"type": "string", "title": "Icon", "default": "\ud83d\udcb0"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}}, "type": "object", "required": ["name", "type"], "title": "CategoryBase"}, "Goal": {"properties": {"id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}, "name": {"type": "string", "title": "Name"}, "target_amount": {"type": "number", "title": "Target Amount"}, "current_amount": {"type": "number", "title": "Current Amount", "default": 0.0}, "deadline": {"type": "string", "format": "date-time", "title": "Deadline"}, "owner_id": {"type": "integer", "title": "Owner Id"}}, "type": "object", "required": ["name", "target_amount", "deadline", "owner_id"], "title": "Goal"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "MessageRequest": {"properties": {"phone": {"type": "string", "title": "Phone"}, "message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["phone", "message"], "title": "MessageRequest"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token"], "title": "Token"}, "TransactionBase": {"properties": {"amount": {"type": "number", "title": "Amount"}, "type": {"type": "string", "title": "Type"}, "description": {"type": "string", "title": "Description"}, "date": {"type": "string", "format": "date-time", "title": "Date"}, "account_id": {"type": "integer", "title": "Account Id"}, "category_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Category Id"}}, "type": "object", "required": ["amount", "type", "description", "account_id"], "title": "TransactionBase"}, "UserCreate": {"properties": {"email": {"type": "string", "title": "Email"}, "full_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Full Name"}, "is_active": {"type": "boolean", "title": "Is Active", "default": true}, "is_superuser": {"type": "boolean", "title": "Is Superuser", "default": false}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "UserCreate"}, "UserOut": {"properties": {"email": {"type": "string", "title": "Email"}, "full_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Full Name"}, "is_active": {"type": "boolean", "title": "Is Active", "default": true}, "is_superuser": {"type": "boolean", "title": "Is Superuser", "default": false}, "id": {"type": "integer", "title": "Id"}}, "type": "object", "required": ["email", "id"], "title": "UserOut"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "login"}}}}}}